# Inline tip5-related functions for a small speedup