Inline a few functions used in tip5 to gain a 1-3 % speedup on the tip5 benchmark.

No inling
tip5/hash_10/Tip5 / Hash 10/10
                        time:   [747.29 ns 752.33 ns 758.48 ns]

tip5/hash_varlen/Tip5 / Hash Variable Length/16384
                        time:   [2.5593 ms 2.5771 ms 2.6001 ms]

tip5/parallel/Tip5 / Parallel Hash/65536
                        time:   [1.4635 ms 1.4834 ms 1.5058 ms]

With inlining
tip5/hash_10/Tip5 / Hash 10/10
                        time:   [722.53 ns 727.50 ns 733.53 ns]

tip5/hash_varlen/Tip5 / Hash Variable Length/16384
                        time:   [2.5468 ms 2.5718 ms 2.5997 ms]

tip5/parallel/Tip5 / Parallel Hash/65536
                        time:   [1.4324 ms 1.4548 ms 1.4830 ms]