@Sword-Smith 2023-01-10
Benchmarks were run on my AMD Ryzen Threadripper PRO 5995WX 64-Cores