@Sword-Smith 2023-01-11
It's worth noting that not all inling gives a speedup. Trial and error is the way forward to figure out where they're suitable.