@Sword-Smith 2023-01-11
On an "11th Gen Intel(R) Core(TM) i9-11950H @ 2.60GHz" I also see a measureable speedup on the order of 3-5 %.

With inlining
```
tip5/hash_10/Tip5 / Hash 10/10                                                                             
                        time:   [754.15 ns 754.82 ns 755.58 ns]
                
tip5/hash_varlen/Tip5 / Hash Variable Length/16384                                                                            
                        time:   [2.3339 ms 2.3347 ms 2.3354 ms]
                
tip5/parallel/Tip5 / Parallel Hash/65536                                                                           
                        time:   [7.9146 ms 7.9516 ms 7.9903 ms]
```
                    
Without inlining
```
tip5/hash_10/Tip5 / Hash 10/10                                                                             
                        time:   [783.66 ns 783.94 ns 784.23 ns]
                
tip5/hash_varlen/Tip5 / Hash Variable Length/16384                                                                            
                        time:   [2.3811 ms 2.3839 ms 2.3870 ms]
                
tip5/parallel/Tip5 / Parallel Hash/65536                                                                           
                        time:   [8.3233 ms 8.3509 ms 8.3787 ms]
```